requirements.txt
python-telegram-bot
python-dotenv
psycopg2
discord-webhook